back: Split the curator's queue out of the verdicts.
`lifecycleController.ts` reached 1122 lines against the guide's "> 800 →
split now", and 354 of those arrived with the curation gate's reader half.
It held one read and three writes that shared the table and nothing else.

The read moves whole: `getReviewQueue` and its seven queries go to
`reviewQueueController.ts`. It has no transaction and no row lock, its
subject is which rows a curator may be *asked* about, and it is the half
that keeps growing — Stage 2 rebuilds it and Stage 3 adds counts to it.

Handing a claimed field back to the source moves too, to
`acceptSourceController.ts`. It answers a different question with a
different answer-holder: `curated_fields` says a person has claimed a
field (ADR-0021), where the two verdicts left behind say whether the
object disappeared and whether this catalogue accepts it — both under the
row's own lock. That third split was left to be decided once the second
was done, and 507 lines against a ~600 target for a controller is what
decided it.

`ACCEPTABLE_FIELDS` and `columnFor` are now their own module, because two
sides need the same answer and neither owns it: the write path builds its
assignment list from it, and the queue uses it to tell a curator which
proposed fields are answerable at all — a card offering "accept" on a
field the endpoint would refuse to write is a button that lies.

Nothing changed but where the text lives, and that is checkable rather
than asserted: `getReviewQueue`, `acceptSourceValue` +
`applyProposedFields`, and the two field helpers are byte-identical to
their versions on main, and the suite passes with no assertion touched —
1321 tests, the same count as before. The only test edits are the two
import paths and the `acceptSourceValue` block moving to sit beside the
handler it covers.

Closes: #526
