+++ backend/src/controllers/experience/acceptSourceController.test.ts
@@ -1,0 +1,296 @@
+/**
+ * Tests for the accept-source path.
+ *
+ * Split from `lifecycleController.test.ts` with the handler it covers (#526).
+ * The rule worth pinning is the one a reader cannot infer from the column names:
+ * accepting a value the source proposed has to release the curator's claim, or
+ * the next run refuses it again — and the proposal is re-resolved under the write
+ * lock, so a claim made since the card was drawn wins.
+ */
+
+import { describe, it, expect, vi, beforeEach } from 'vitest';
+
+vi.mock('../../db/index.js', () => ({
+  pool: { query: vi.fn(), connect: vi.fn() },
+  // Mirrors the real one, returning the rollback's own failure — which is
+  // what `client.release()` needs to destroy a client carrying an open
+  // transaction. A stub that swallowed it would make that untestable.
+  rollbackQuietly: async (c: { query: (s: string) => unknown }) => {
+    try { await c.query('ROLLBACK'); return undefined; } catch (e) { return e as Error; }
+  },
+}));
+
+import { pool } from '../../db/index.js';
+import { acceptSourceValue } from './acceptSourceController.js';
+import { CHANGESET_LANDED_SQL } from '../../services/sync/syncLogMarkers.js';
+
+const mockedQuery = pool.query as unknown as ReturnType<typeof vi.fn>;
+const mockedConnect = pool.connect as unknown as ReturnType<typeof vi.fn>;
+
+function makeRes() {
+  return { json: vi.fn(), status: vi.fn().mockReturnThis() };
+}
+
+const CURATOR = { id: 7, role: 'curator' as const };
+const ADMIN = { id: 1, role: 'admin' as const };
+
+/**
+ * Captures what the transaction ran, so assertions can read the statements.
+ *
+ * `claimed` answers the `FOR UPDATE` re-read, which is the point of the mock:
+ * this handler reads `curated_fields` inside its own transaction rather than
+ * trusting a value fetched before the lock, so a claim made since the card was
+ * drawn has to be able to win. That is the only column the locked read selects
+ * here — the lifecycle axes belong to the verdict handlers and their mock, in
+ * `lifecycleController.test.ts`.
+ */
+function makeClient(claimed?: string[], proposal?: unknown[]) {
+  const queries: Array<{ sql: string; params: unknown[] }> = [];
+  return {
+    queries,
+    client: {
+      query: vi.fn(async (sql: string, params?: unknown[]) => {
+        queries.push({ sql, params: params ?? [] });
+        if (sql.includes('experience_sync_changes')) return { rows: proposal ?? [] };
+        if (sql.includes('FOR UPDATE')) return { rows: [{ curated_fields: claimed ?? [] }] };
+        return { rows: [] };
+      }),
+      release: vi.fn(),
+    },
+  };
+}
+
+describe('acceptSourceValue', () => {
+  beforeEach(() => {
+    mockedQuery.mockReset();
+    mockedConnect.mockReset();
+  });
+
+  it('refuses a curator whose scope does not reach the experience', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    mockedQuery.mockResolvedValueOnce({ rows: [{ unrestricted: false, scoped_region_id: null }] });
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: CURATOR, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 9 } } as never,
+      res as never,
+    );
+
+    expect(res.status).toHaveBeenCalledWith(403);
+  });
+
+  it('says so when the source never proposed anything', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    mockedConnect.mockResolvedValue(makeClient(['name'], []).client);
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 9 } } as never,
+      res as never,
+    );
+
+    expect(res.status).toHaveBeenCalledWith(409);
+  });
+
+  it('releases the curator claim, or the next run would refuse the value again', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 9, changed_fields: [{ field: 'name', new: 'Renamed upstream', curatedConflict: true }] }];
+    const { client, queries } = makeClient(['name', 'description'], PROPOSAL);
+    mockedConnect.mockResolvedValue(client);
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 9 } } as never,
+      res as never,
+    );
+
+    const update = queries.find(q => q.sql.includes('UPDATE experiences'));
+    expect(update?.sql).toContain('name = $2');
+    expect(update?.params).toContain('Renamed upstream');
+    // 'name' released, 'description' still claimed
+    expect(update?.params).toContain(JSON.stringify(['description']));
+    expect(res.json).toHaveBeenCalledWith(expect.objectContaining({ fromSyncLogId: 9 }));
+  });
+
+  it('will not write a proposal the source has since withdrawn', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    // The withdrawal check is in the SQL, so a withdrawn proposal comes back
+    // as no row — the same answer as no proposal ever existing
+    const { client, queries } = makeClient(['name'], []);
+    mockedConnect.mockResolvedValue(client);
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 41 } } as never,
+      res as never,
+    );
+
+    const lookup = queries.find(q => q.sql.includes('experience_sync_changes'));
+    expect(lookup?.sql).toContain('last_seen_sync_log_id');
+    // Same gate as the queue, or a card fetched before a run started would be
+    // refused mid-run with "no proposal on record", which is false
+    // The two copies must stay identical, or the endpoint would refuse a
+    // newer proposal while writing a retracted one. Sharing one definition is
+    // what makes that structural; this asserts they really do share it.
+    const queueSql = CHANGESET_LANDED_SQL.replace(/\s+/g, ' ').trim();
+    expect(lookup!.sql.replace(/\s+/g, ' ')).toContain(queueSql);
+    expect(res.status).toHaveBeenCalledWith(409);
+    expect(queries.some(q => q.sql.includes('UPDATE experiences'))).toBe(false);
+  });
+
+  it('resolves the proposal under the same lock that writes it', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 9, changed_fields: [{ field: 'name', new: 'X', curatedConflict: true }] }];
+    const { client, queries } = makeClient(['name'], PROPOSAL);
+    mockedConnect.mockResolvedValue(client);
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 9 } } as never,
+      makeRes() as never,
+    );
+
+    // Resolving it before the lock would let a run landing in between have its
+    // values written under the run id the curator sent — the substitution
+    // expectedSyncLogId exists to refuse
+    expect(queries[0].sql).toBe('BEGIN');
+    expect(queries.some(q => q.sql.includes('experience_sync_changes'))).toBe(true);
+    const lookup = queries.findIndex(q => q.sql.includes('experience_sync_changes'));
+    const write = queries.findIndex(q => q.sql.includes('UPDATE experiences'));
+    expect(lookup).toBeGreaterThan(0);
+    expect(write).toBeGreaterThan(lookup);
+  });
+
+  it('finishes rolling back before letting go of the client', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 42, changed_fields: [{ field: 'name', new: 'X', curatedConflict: true }] }];
+    const { client, queries } = makeClient(['name'], PROPOSAL);
+    let rollbackDone = false;
+    let releasedBeforeRollback = false;
+    const inner = client.query.getMockImplementation()!;
+    client.query.mockImplementation(async (sql: string, params?: unknown[]) => {
+      if (sql === 'ROLLBACK') {
+        await new Promise(r => setTimeout(r, 5));
+        rollbackDone = true;
+        queries.push({ sql, params: params ?? [] });
+        return { rows: [] };
+      }
+      return inner(sql, params);
+    });
+    client.release.mockImplementation(() => { if (!rollbackDone) releasedBeforeRollback = true; });
+    mockedConnect.mockResolvedValue(client);
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 41 } } as never,
+      makeRes() as never,
+    );
+
+    // An unawaited refusal settles the try block at once, so `finally` hands
+    // the client back while its ROLLBACK is still running on the wire
+    expect(releasedBeforeRollback).toBe(false);
+  });
+
+  it('refuses a proposal a newer run has replaced', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    // The card was drawn from run 41; run 42 has since proposed something else
+    const PROPOSAL = [{ sync_log_id: 42, changed_fields: [{ field: 'name', new: 'Rathaus', curatedConflict: true }] }];
+    const { client, queries } = makeClient(['name'], PROPOSAL);
+    mockedConnect.mockResolvedValue(client);
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 41 } } as never,
+      res as never,
+    );
+
+    // Writing here would replace the curator's edit with a value they never
+    // saw — the same exposure `expected` closes on the verdict path
+    expect(res.status).toHaveBeenCalledWith(409);
+    expect(queries.some(q => q.sql.includes('UPDATE experiences'))).toBe(false);
+  });
+
+  it('names the run its value came from, since a later one may differ', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 42, changed_fields: [{ field: 'name', new: 'X', curatedConflict: true }] }];
+    mockedConnect.mockResolvedValue(makeClient(['name'], PROPOSAL).client);
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 42 } } as never,
+      res as never,
+    );
+
+    expect(res.json).toHaveBeenCalledWith(expect.objectContaining({ applied: ['name'], released: [], fromSyncLogId: 42 }));
+  });
+
+  it('releases a field it cannot write, since nothing else ever would', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 9, changed_fields: [{ field: 'tags', new: ['a'], curatedConflict: true }] }];
+    const { client, queries } = makeClient(['tags', 'name'], PROPOSAL);
+    mockedConnect.mockResolvedValue(client);
+    const res = makeRes();
+
+    // 'tags' carries a real proposal but has no column here. Editing would not
+    // clear it either — every other writer of curated_fields only ever adds —
+    // so releasing the claim is the only way off the queue, and the next run
+    // then writes the value through the ordinary upsert.
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['tags'], expectedSyncLogId: 9 } } as never,
+      res as never,
+    );
+
+    const update = queries.find(q => q.sql.includes('UPDATE experiences'));
+    expect(update?.sql).not.toContain('tags =');
+    expect(update?.params).toContain(JSON.stringify(['name']));
+    expect(res.json).toHaveBeenCalledWith(expect.objectContaining({ applied: [], released: ['tags'] }));
+  });
+
+  it('reads the claim it is about to rewrite under the same lock', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 9, changed_fields: [{ field: 'name', new: 'X', curatedConflict: true }] }];
+    const { client, queries } = makeClient(['name'], PROPOSAL);
+    mockedConnect.mockResolvedValue(client);
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 9 } } as never,
+      makeRes() as never,
+    );
+
+    // Reading it before the transaction and writing the filtered result back
+    // would drop whatever a concurrent edit claimed in between
+    const order = queries.map(q => q.sql.trim().split(/\s+/).slice(0, 2).join(' '));
+    expect(order[0]).toBe('BEGIN');
+    expect(queries[1].sql).toContain('FOR UPDATE');
+    expect(queries.findIndex(q => q.sql.includes('UPDATE experiences'))).toBeGreaterThan(1);
+  });
+
+  it('does not overwrite an answer another curator already gave', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 9, changed_fields: [{ field: 'name', new: 'X', curatedConflict: true }] }];
+    // The claim was released while this request was in flight
+    const { client, queries } = makeClient([], PROPOSAL);
+    mockedConnect.mockResolvedValue(client);
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['name'], expectedSyncLogId: 9 } } as never,
+      res as never,
+    );
+
+    expect(queries.some(q => q.sql.includes('UPDATE experiences'))).toBe(false);
+    expect(res.status).toHaveBeenCalledWith(409);
+  });
+
+  it('refuses a field the source did not propose', async () => {
+    mockedQuery.mockResolvedValueOnce({ rows: [{ id: 5, category_id: 1 }] });
+    const PROPOSAL = [{ sync_log_id: 9, changed_fields: [{ field: 'name', new: 'X', curatedConflict: true }] }];
+    mockedConnect.mockResolvedValue(makeClient(['name'], PROPOSAL).client);
+    const res = makeRes();
+
+    await acceptSourceValue(
+      { user: ADMIN, params: { id: '5' }, body: { fields: ['description'], expectedSyncLogId: 9 } } as never,
+      res as never,
+    );
+
+    expect(res.status).toHaveBeenCalledWith(409);
+  });
+});
