+++ backend/src/controllers/experience/index.ts
@@ -8,5 +8,7 @@ export * from './experienceLocationController.js';
 export * from './experienceTreasureController.js';
 export * from './curationController.js';
 export * from './lifecycleController.js';
+export * from './reviewQueueController.js';
+export * from './acceptSourceController.js';
 export * from './publishController.js';
 export * from './experienceNewBadge.js';
